% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artificialcarbontax.R
\docType{data}
\name{artificialcarbontax}
\alias{artificialcarbontax}
\title{The artificialcarbontax dataset}
\format{A data frame with 10 observations of 3 variables:
\describe{
 \item{FROM}{FROM, date of starting tax rate}
 \item{TO}{TO, date of ending tax rate. The last row is considered the present day rate indicated by the package creation date.}
 \item{RATE}{RATE, in dollars}
 }}
\source{
\url{http://catalog.data.gov/dataset/historical-tax-rates}
}
\usage{
artificialcarbontax
}
\description{
This dataset contains a list of historical tax rates for
 artificial carbon per wine gallon from TTB.
}
\keyword{datasets}

