% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beermaterials.R
\docType{data}
\name{beermaterials}
\alias{beermaterials}
\title{The beermaterials dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 120 rows and 11 columns.}
\source{
\url{https://catalog.data.gov/dataset/beer-production-and-operations-reports}
}
\usage{
beermaterials
}
\description{
The monthly aggregation of materials used in US Breweries as reported by the TTB.
    The dataset was transcribed from PDF's to a .rda dataset.
}
\keyword{datasets}

