% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirittax.R
\docType{data}
\name{spirittax}
\alias{spirittax}
\title{The spirittax dataset}
\format{A data frame with 15 observations of 3 variables:
\describe{
 \item{FROM}{FROM, date of starting tax rate}
 \item{TO}{TO, date of ending tax rate. The last row is considered the present day rate indicated by the package creation date.}
 \item{RATE}{RATE, in dollars}
 }}
\source{
\url{http://catalog.data.gov/dataset/historical-tax-rates}
}
\usage{
spirittax
}
\description{
This dataset contains a list of historical tax rates for
distilled spirits per proof gallon from TTB.
}
\keyword{datasets}

