% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobaccotax.R
\docType{data}
\name{tobaccotax}
\alias{tobaccotax}
\title{The tobaccotax dataset}
\format{A data frame with 46 observations of 3 variables:
\describe{
 \item{ITEM}{ITEM, specific tobacco-based product.}
 \item{FROM}{FROM, date of starting tax rate. The first date does not have a listed day, so 09/01/1862 is used.}
 \item{TO}{TO, date of ending tax rate. The last row is considered the present day rate indicated by the package creation date.}
 \item{RATE}{RATE, in dollars. Small cigarettes through Large cigars are per thousand. Chewing tobacco through Roll your own are per pound. Cigarette papers and tubes are per 50.}
 }}
\source{
\url{http://catalog.data.gov/dataset/historical-tax-rates}
}
\usage{
tobaccotax
}
\description{
This dataset contains a list of historical tax rates for
 tobacco and tobacco products from TTB.
}
\keyword{datasets}

