% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttbbeer.R
\docType{package}
\name{ttbbeer}
\alias{ttbbeer}
\alias{ttbbeer-package}
\title{ttbbeer: A package for US Brewery data sets.}
\description{
The ttbbeer package includes at current state, one data set of
   a collection of 10 years (2006 - 2015) worth of data on materials
   used at U.S. breweries in pounds reported by the Brewer's Report
   of Operations and the Quarterly Brewer's Report of Operations
   forms, ready for data analysis.
}
\section{ttbbeer functions}{

  The ttbbeer functions include data("beermaterials").
}

