% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winetax1421.R
\docType{data}
\name{winetax1421}
\alias{winetax1421}
\title{The winetax1421 dataset}
\format{A data frame with 11 observations of 3 variables:
\describe{
 \item{FROM}{FROM, date of starting tax rate}
 \item{TO}{TO, date of ending tax rate. The last row is considered the present day rate indicated by the package creation date.}
 \item{RATE}{RATE, in dollars}
 }}
\source{
\url{http://catalog.data.gov/dataset/historical-tax-rates}
}
\usage{
winetax1421
}
\description{
This dataset contains a list of historical tax rates for
wine 14 - 21% ABV per wine gallon from TTB.
}
\keyword{datasets}

