% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beermaterials.R
\docType{data}
\name{beermaterials}
\alias{beermaterials}
\title{The beermaterials dataset
The dataset was transcribed from PDF's to a .rda dataset}
\format{A data frame with 120 observations of 11 variables:
\describe{
 \item{Month}{Month of the year}
 \item{Year}{Calendar Year}
 \item{Malt_and_malt_products}{Malt and malt products, in pounds}
 \item{Corn_and _corn_products}{Corn and corn products, in pounds}
 \item{Rice_and_rice_products}{Rice and rice products, in pounds}
 \item{Barley_and_barley_products}{Barley and barley products, in pounds}
 \item{Wheat_and_wheat_products}{Wheat and wheat products}
 \item{Sugar_and_syrups}{Sugar and syrups, in pounds}
 \item{Hops_dry}{Hops in dry form, in pounds}
 \item{Hops_extracts}{Hops in extracted form, in pounds}
 \item{Other}{Other materials such as flavor, in pounds}
}}
\source{
\url{https://catalog.data.gov/dataset/beer-production-and-operations-reports}
}
\usage{
beermaterials
}
\description{
The beermaterials dataset
The dataset was transcribed from PDF's to a .rda dataset
}
\keyword{datasets}

